+++ examples/rayon-compare/build.rs
@@ -1,3 +1,0 @@
-fn main() {
-    lace_native_build::process("tasks.def").compile();
-}